Limop
App

+++ Limop/app/src/main/java/br/com/limopestoques/limop/EditarProdVenda.java
@@ -433,7 +433,7 @@ public class EditarProdVenda extends AppCompatActivity {
                 || desconto.getText().length() == 0  || vencimento.getText().length() == 0) {
             Toast.makeText(this, getString(R.string.preenchaoscamposcorretamente), Toast.LENGTH_SHORT).show();
         }else{
-            Integer Desconto = Integer.parseInt(desconto.getText().toString());
+            Double Desconto = Double.parseDouble(desconto.getText().toString());
             if(Desconto > 100){
                 Toast.makeText(this, "Porcentagem inválida!", Toast.LENGTH_SHORT).show();
             }else{
